Reject promises for workers that are about to be killed
Currently the promises for workers that are killed are left
dangling which is problematic because dangling promises
can result in memory leaks and other bad outcomes.
A simple way to avoid that is to reject the promise
with a specific AbortError that the user can catch
and handle as they please. This is also the approach
that axios is using for example for the same case
if you kill a request before it returns a response.