fix: mark worker as not running immediately after being killed (#171)
This change is necessary so that the client could kill a worker
and then immediately start a new worker. Without this change
`useWorker()` will error out and say that you can only run one
instance at a time because the running status is not updated
until the next render cycle. If we mark it as not running
immediately then the client can immediately create a new
worker without having to wait for that to happen.